Remove original native_gpu_future and replace with its non-allocating cousin
