+++ src/backend/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod native_gpu_future;
-pub(crate) mod no_alloc_future;
